+++ client/src/routes/api/home.js
